Rm .helix
The .helix was there to help the lsp deal with the
front and backend files. The thing is, that does
not need to be done. The front and backend are two
different "projects", it just so happens that both
are written in rust. The lsp works fine if you
open one project at a time