Bug fixes
The graph power limits adjustment was off. With the
system normalized, the numbers just didn't match.
Now we take into consideration the normalization
to calculate the limits the backend sees.

Also, the graph limits took into consideration the
entire history of spectra. Now it scales up to
only the living one (frozen + active)